+++ packages/git-mob-core/src/git-mob-api/git-config.ts
@@ -7,3 +7,11 @@ export async function getLocalCommitTemplate() {
 export async function getGlobalCommitTemplate() {
   return getConfig('--global commit.template');
 }
+
+export async function getGitUserName() {
+  return getConfig('user.name');
+}
+
+export async function getGitUserEmail() {
+  return getConfig('user.email');
+}